requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-huggingface
faiss-cpu
python-multipart